requirements.txt
Django==1.5.1
Markdown==2.3.1
Pillow==1.7.8
South==0.7.6
Unipath==1.0
django-allauth==0.12.0
django-braces==1.1.0
django-gravatar==0.1.0
django-selectable==0.7.0
djrill==0.3.0
httplib2==0.8
logilab-astng==0.24.3
logilab-common==0.59.1
micawber==0.2.6
model-mommy==1.0
oauth2==1.5.211
oauthlib==0.5.0
python-dateutil==2.1
python-openid==2.2.5
requests==1.2.3
requests-oauthlib==0.3.2
simplejson==3.1.2
six==1.3.0
